--- components/separator.tsx
@@ -1,3 +1,0 @@
-export default function Separator() {
-  return <div className="h-0.5 bg-muted my-4" />;
-}
